Ajout d'une garde si un glob échoue
git-svn-id: http://afi-forge.afi-sa.fr/svn/opacce/trunk@24 e3cc70dd-a52f-4065-8a26-0e09943c8c5c

+++ application/modules/opac/controllers/IndexController.php
@@ -161,6 +161,9 @@ class IndexController extends Zend_Controller_Action {
 				mkdir(PATH_CAPTCHA);
 
 		$pngs = glob( PATH_CAPTCHA.'*.png' );
+		if (!is_array($pngs))
+			return;
+
 		foreach ($pngs as $png) {
 			if ((time() - filemtime($png )) > 60 )
 				unlink($png);
